api/requirements.txt
pydub==0.23.1
pytube==9.5.2
google-cloud==0.34.0
google-cloud-speech==1.2.0
google-cloud-storage==1.20.0
google-resumable-media==0.4.1